+++ core/src/builtin_widgets/theme/palette.rs
@@ -14,7 +14,7 @@ use super::*;
 ///
 /// let _w = Palette::default().with_child(fn_widget! {
 ///   // Every widget created in this scope can access the `Palette`.
-///   let _primary = Palette::of(ctx!()).primary();
+///   let _primary = Palette::of(BuildCtx::get()).primary();
 ///   Void
 /// });
 /// ```